moved Sensor type to Util
